Infrastructure enhancements
### What's done:
* Added grafana.ini
* Updated docker-compose.yaml.template
* Extracted DatabaseTestBase to a junit5 extension
