+++ save-backend/src/test/kotlin/org/cqfn/save/backend/SaveResultTest.kt
@@ -1,13 +1,14 @@
 package org.cqfn.save.backend
 
 import org.cqfn.save.backend.repository.TestExecutionRepository
-import org.cqfn.save.backend.utils.DatabaseTestBase
+import org.cqfn.save.backend.utils.MySqlExtension
 import org.cqfn.save.domain.TestResultStatus
 import org.cqfn.save.entities.TestExecution
 
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient
 import org.springframework.boot.test.context.SpringBootTest
@@ -19,7 +20,8 @@ import java.time.LocalDateTime
 
 @SpringBootTest(classes = [SaveApplication::class])
 @AutoConfigureWebTestClient
-class SaveResultTest : DatabaseTestBase() {
+@ExtendWith(MySqlExtension::class)
+class SaveResultTest {
     private val date = LocalDateTime.now().withNano(0)
 
     @Autowired
