+++ save-deploy/grafana.ini
@@ -1,0 +1,3 @@
+[server]
+root_url=%(protocol)s://localhost:%(http_port)s/grafana
+serve_from_sub_path = true
